requirements.txt
asgiref
asyncio
attrs
colorama
Django
django-cors-headers
django-environ
django-extensions
django-redis
django-prometheus
django-ratelimit
python-json-logger
django-silk
djangorestframework
djangorestframework-simplejwt
drf-spectacular
drf-spectacular-sidecar
inflection
iniconfig
jsonschema
jsonschema-specifications
mysql
mysqlclient
packaging
pluggy
psycopg2
PyJWT
pytest
python-dotenv
PyYAML
redis
referencing
rpds-py
sqlparse
tzdata
uritemplate
wheel
whitenoise
gunicorn
